fix: add ui components and account store
