+++ src/components/ErrorMessage/errorMessage.css
@@ -1,0 +1,7 @@
+.error-message {
+    background-color: rgb(239, 45, 45);
+    height: 100px;
+    width: 400px;
+    color: white;
+    font-size: x-large;
+}
