+++ src/components/LoadButton/loadButton.css
@@ -1,0 +1,7 @@
+.load-button {
+    border: 0;
+    padding: 8px;
+    background-color: rgb(129, 152, 226);
+    color: #f4f1f1;
+    font-weight: bolder;
+}
