+++ src/components/TextMessage/TextMessage.css
@@ -1,0 +1,3 @@
+.text-message {
+    min-height: 180px;
+}
