--- src/components/error-message/error-message.css
@@ -1,7 +1,0 @@
-.error-message {
-    background-color: rgb(239, 45, 45);
-    height: 100px;
-    width: 400px;
-    color: white;
-    font-size: x-large;
-}
