--- src/components/text-message/text-message.css
@@ -1,6 +1,0 @@
-.text-message {
-    background-color: green;
-    min-height: 300px;
-    width: 400px;
-    color: white;
-}
