+++ src/main.tsx
@@ -1,7 +1,11 @@
 import { createRoot } from 'react-dom/client'
 import 'bootstrap/dist/css/bootstrap.min.css';
 import App from './App.tsx'
+import { Provider } from 'react-redux';
+import { store } from './store/store.ts';
 
 createRoot(document.getElementById('root')!).render(
-    <App />
+    <Provider store={store}>
+        <App />
+    </Provider>
 )
