+++ src/model/Account.ts
@@ -1,0 +1,3 @@
+export interface Account {
+    email: string
+}
