+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,0 +1,14 @@
+import { type FC } from 'react';
+
+/*
+    Страница не найдена (404)
+*/
+const NotFoundPage: FC = () => {
+    return (
+        <>
+            <h2>Страница не найдена</h2>
+        </>
+    )
+}
+
+export default NotFoundPage;
