--- src/pages/login-page/login-page.tsx
@@ -1,15 +1,0 @@
-import { type FC } from 'react';
-// import './login-page.css';
-
-/*
-    Страница аутентификации
-*/
-const LoginPage: FC = () => {
-    return (
-        <>
-            <h2>Страница аутентификации</h2>
-        </>
-    )
-}
-
-export default LoginPage;
