--- src/pages/notfound-page/notfound-page.tsx
@@ -1,15 +1,0 @@
-import { type FC } from 'react';
-// import './notfound-page.css';
-
-/*
-    Страница не найдена (404)
-*/
-const NotFoundPage: FC = () => {
-    return (
-        <>
-            <h2>Страница не найдена</h2>
-        </>
-    )
-}
-
-export default NotFoundPage;
