--- src/pages/register-page/register-page.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-// import './register-page.css';
-
-/*
-    Страница регистрации
-*/
-const RegisterPage: React.FC = () => {
-    return (
-        <>
-            <h2>Страница регистрации</h2>
-        </>
-    )
-}
-
-export default RegisterPage;
