+++ src/store/store.ts
@@ -1,8 +1,9 @@
 import { configureStore } from "@reduxjs/toolkit";
+import { account } from "./account";
 
 export const store = configureStore({
     reducer: {
-        // append reducers
+        account
     }
 })
